Changed ChooseRacePanel definitions to model/view/controller pattern.
