+++ OpenTESArena/src/Interface/CharacterCreationUiController.cpp
@@ -3,6 +3,7 @@
 #include "CharacterCreationUiController.h"
 #include "CharacterCreationUiModel.h"
 #include "CharacterCreationUiView.h"
+#include "ChooseAttributesPanel.h"
 #include "ChooseClassCreationPanel.h"
 #include "ChooseClassPanel.h"
 #include "ChooseGenderPanel.h"
@@ -123,6 +124,399 @@ void CharacterCreationUiController::onChooseNameAcceptButtonSelected(Game &game,
 	charCreationState.setName(acceptedName.c_str());
 
 	game.setPanel<ChooseGenderPanel>(game);
+}
+
+void CharacterCreationUiController::onBackToChooseGenderButtonSelected(Game &game)
+{
+	game.setPanel<ChooseGenderPanel>(game);
+}
+
+void CharacterCreationUiController::onChooseRaceInitialPopUpButtonSelected(Game &game)
+{
+	game.popSubPanel();
+}
+
+void CharacterCreationUiController::onChooseRaceProvinceButtonSelected(Game &game, int raceID)
+{
+	// Set character creation race.
+	auto &charCreationState = game.getCharacterCreationState();
+	charCreationState.setRaceIndex(raceID);
+
+	// Generate the race selection message box.
+	auto &textureManager = game.getTextureManager();
+	auto &renderer = game.getRenderer();
+
+	const Color textColor(52, 24, 8);
+
+	MessageBoxSubPanel::Title messageBoxTitle;
+	messageBoxTitle.textBox = [&game, &renderer, &textColor]()
+	{
+		const int lineSpacing = 1;
+		const auto &fontLibrary = game.getFontLibrary();
+		const RichTextString richText(
+			CharacterCreationUiModel::getChooseRaceProvinceConfirmTitleText(game),
+			FontName::A,
+			textColor,
+			TextAlignment::Center,
+			lineSpacing,
+			fontLibrary);
+
+		const Int2 center(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2),
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) - 22);
+
+		return std::make_unique<TextBox>(center, richText, fontLibrary, renderer);
+	}();
+
+	messageBoxTitle.texture = [&textureManager, &renderer, &messageBoxTitle]()
+	{
+		const int width = messageBoxTitle.textBox->getRect().getWidth() + 22;
+		const int height = 60;
+		return TextureUtils::generate(TextureUtils::PatternType::Parchment, width, height,
+			textureManager, renderer);
+	}();
+
+	messageBoxTitle.textureX = (ArenaRenderUtils::SCREEN_WIDTH / 2) -
+		(messageBoxTitle.texture.getWidth() / 2) - 1;
+	messageBoxTitle.textureY = (ArenaRenderUtils::SCREEN_HEIGHT / 2) -
+		(messageBoxTitle.texture.getHeight() / 2) - 21;
+
+	MessageBoxSubPanel::Element messageBoxYes;
+	messageBoxYes.textBox = [&game, &renderer, &textColor]()
+	{
+		const auto &fontLibrary = game.getFontLibrary();
+		const RichTextString richText(
+			"Yes",
+			FontName::A,
+			textColor,
+			TextAlignment::Center,
+			fontLibrary);
+
+		const Int2 center(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2) - 1,
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) + 28);
+
+		return std::make_unique<TextBox>(center, richText, fontLibrary, renderer);
+	}();
+
+	messageBoxYes.texture = [&textureManager, &renderer, &messageBoxTitle]()
+	{
+		const int width = messageBoxTitle.texture.getWidth();
+		return TextureUtils::generate(TextureUtils::PatternType::Parchment, width, 40,
+			textureManager, renderer);
+	}();
+
+	messageBoxYes.function = [raceID](Game &game)
+	{
+		CharacterCreationUiController::onChooseRaceProvinceConfirmButtonSelected(game, raceID);
+	};
+
+	messageBoxYes.textureX = messageBoxTitle.textureX;
+	messageBoxYes.textureY = messageBoxTitle.textureY + messageBoxTitle.texture.getHeight();
+
+	MessageBoxSubPanel::Element messageBoxNo;
+	messageBoxNo.textBox = [&game, &renderer, &textColor]()
+	{
+		const auto &fontLibrary = game.getFontLibrary();
+		const RichTextString richText(
+			"No",
+			FontName::A,
+			textColor,
+			TextAlignment::Center,
+			fontLibrary);
+
+		const Int2 center(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2) - 1,
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) + 68);
+
+		return std::make_unique<TextBox>(center, richText, fontLibrary, renderer);
+	}();
+
+	messageBoxNo.texture = [&textureManager, &renderer, &messageBoxYes]()
+	{
+		const int width = messageBoxYes.texture.getWidth();
+		const int height = messageBoxYes.texture.getHeight();
+		return TextureUtils::generate(TextureUtils::PatternType::Parchment, width, height,
+			textureManager, renderer);
+	}();
+
+	messageBoxNo.function = CharacterCreationUiController::onChooseRaceProvinceCancelButtonSelected;
+	messageBoxNo.textureX = messageBoxYes.textureX;
+	messageBoxNo.textureY = messageBoxYes.textureY + messageBoxYes.texture.getHeight();
+
+	auto cancelFunction = messageBoxNo.function;
+
+	std::vector<MessageBoxSubPanel::Element> messageBoxElements;
+	messageBoxElements.push_back(std::move(messageBoxYes));
+	messageBoxElements.push_back(std::move(messageBoxNo));
+
+	auto messageBox = std::make_unique<MessageBoxSubPanel>(
+		game, std::move(messageBoxTitle), std::move(messageBoxElements),
+		cancelFunction);
+
+	game.pushSubPanel(std::move(messageBox));
+}
+
+void CharacterCreationUiController::onChooseRaceProvinceConfirmButtonSelected(Game &game, int raceID)
+{
+	game.popSubPanel();
+
+	const Color textColor(48, 12, 12);
+
+	// Generate all of the parchments leading up to the attributes panel,
+	// and link them together so they appear after each other.
+	auto toAttributes = [](Game &game)
+	{
+		game.popSubPanel();
+		game.setPanel<ChooseAttributesPanel>(game);
+	};
+
+	auto toFourthSubPanel = [textColor, toAttributes](Game &game)
+	{
+		game.popSubPanel();
+
+		const Int2 center((ArenaRenderUtils::SCREEN_WIDTH / 2) - 1, 98);
+
+		const std::string text = [&game]()
+		{
+			const auto &exeData = game.getBinaryAssetLibrary().getExeData();
+			std::string segment = exeData.charCreation.confirmedRace4;
+			segment = String::replace(segment, '\r', '\n');
+
+			return segment;
+		}();
+
+		const int lineSpacing = 1;
+
+		const RichTextString richText(
+			text,
+			FontName::Arena,
+			textColor,
+			TextAlignment::Center,
+			lineSpacing,
+			game.getFontLibrary());
+
+		const int textureHeight = std::max(richText.getDimensions().y + 8, 40);
+		Texture texture = TextureUtils::generate(TextureUtils::PatternType::Parchment,
+			richText.getDimensions().x + 20, textureHeight,
+			game.getTextureManager(), game.getRenderer());
+
+		const Int2 textureCenter(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2) - 1,
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) - 1);
+
+		auto fourthSubPanel = std::make_unique<TextSubPanel>(
+			game, center, richText, toAttributes, std::move(texture),
+			textureCenter);
+
+		game.pushSubPanel(std::move(fourthSubPanel));
+	};
+
+	auto toThirdSubPanel = [textColor, toFourthSubPanel](Game &game)
+	{
+		game.popSubPanel();
+
+		const Int2 center((ArenaRenderUtils::SCREEN_WIDTH / 2) - 1, 98);
+
+		const std::string text = [&game]()
+		{
+			const auto &binaryAssetLibrary = game.getBinaryAssetLibrary();
+			const auto &exeData = binaryAssetLibrary.getExeData();
+			std::string segment = exeData.charCreation.confirmedRace3;
+			segment = String::replace(segment, '\r', '\n');
+
+			const auto &charCreationState = game.getCharacterCreationState();
+			const auto &charClassLibrary = game.getCharacterClassLibrary();
+			const int charClassDefID = charCreationState.getClassDefID();
+			const auto &charClassDef = charClassLibrary.getDefinition(charClassDefID);
+
+			const auto &preferredAttributes = exeData.charClasses.preferredAttributes;
+			DebugAssertIndex(preferredAttributes, charClassDefID);
+			const std::string &preferredAttributesStr = preferredAttributes[charClassDefID];
+
+			// Replace first %s with desired class attributes.
+			size_t index = segment.find("%s");
+			segment.replace(index, 2, preferredAttributesStr);
+
+			// Replace second %s with class name.
+			index = segment.find("%s");
+			segment.replace(index, 2, charClassDef.getName());
+
+			return segment;
+		}();
+
+		const int lineSpacing = 1;
+
+		const RichTextString richText(
+			text,
+			FontName::Arena,
+			textColor,
+			TextAlignment::Center,
+			lineSpacing,
+			game.getFontLibrary());
+
+		const int textureHeight = std::max(richText.getDimensions().y + 18, 40);
+		Texture texture = TextureUtils::generate(TextureUtils::PatternType::Parchment,
+			richText.getDimensions().x + 20, textureHeight,
+			game.getTextureManager(), game.getRenderer());
+
+		const Int2 textureCenter(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2) - 1,
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) - 1);
+
+		auto thirdSubPanel = std::make_unique<TextSubPanel>(
+			game, center, richText, toFourthSubPanel, std::move(texture),
+			textureCenter);
+
+		game.pushSubPanel(std::move(thirdSubPanel));
+	};
+
+	auto toSecondSubPanel = [textColor, toThirdSubPanel](Game &game)
+	{
+		game.popSubPanel();
+
+		const Int2 center((ArenaRenderUtils::SCREEN_WIDTH / 2) - 1, 98);
+
+		const std::string text = [&game]()
+		{
+			const auto &exeData = game.getBinaryAssetLibrary().getExeData();
+			std::string segment = exeData.charCreation.confirmedRace2;
+			segment = String::replace(segment, '\r', '\n');
+
+			const auto &charCreationState = game.getCharacterCreationState();
+			const int raceIndex = charCreationState.getRaceIndex();
+
+			// Get race description from TEMPLATE.DAT.
+			const auto &templateDat = game.getTextAssetLibrary().getTemplateDat();
+			constexpr std::array<int, 8> raceTemplateIDs =
+			{
+				1409, 1410, 1411, 1412, 1413, 1414, 1415, 1416
+			};
+
+			DebugAssertIndex(raceTemplateIDs, raceIndex);
+			const auto &entry = templateDat.getEntry(raceTemplateIDs[raceIndex]);
+			std::string raceDescription = entry.values.front();
+
+			// Re-distribute newlines at 40 character limit.
+			raceDescription = String::distributeNewlines(raceDescription, 40);
+
+			// Append race description to text segment.
+			segment += "\n" + raceDescription;
+
+			return segment;
+		}();
+
+		const int lineSpacing = 1;
+
+		const RichTextString richText(
+			text,
+			FontName::Arena,
+			textColor,
+			TextAlignment::Center,
+			lineSpacing,
+			game.getFontLibrary());
+
+		const int textureHeight = std::max(richText.getDimensions().y + 14, 40);
+		Texture texture = TextureUtils::generate(TextureUtils::PatternType::Parchment,
+			richText.getDimensions().x + 20, textureHeight,
+			game.getTextureManager(), game.getRenderer());
+
+		const Int2 textureCenter(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2) - 1,
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) - 1);
+
+		auto secondSubPanel = std::make_unique<TextSubPanel>(
+			game, center, richText, toThirdSubPanel, std::move(texture),
+			textureCenter);
+
+		game.pushSubPanel(std::move(secondSubPanel));
+	};
+
+	std::unique_ptr<Panel> firstSubPanel = [&game, &textColor, toSecondSubPanel]()
+	{
+		const Int2 center((ArenaRenderUtils::SCREEN_WIDTH / 2) - 1, 98);
+
+		const std::string text = [&game]()
+		{
+			const auto &binaryAssetLibrary = game.getBinaryAssetLibrary();
+			const auto &exeData = binaryAssetLibrary.getExeData();
+			std::string segment = exeData.charCreation.confirmedRace1;
+			segment = String::replace(segment, '\r', '\n');
+
+			const auto &charCreationState = game.getCharacterCreationState();
+			const int raceIndex = charCreationState.getRaceIndex();
+
+			const auto &charCreationProvinceNames = exeData.locations.charCreationProvinceNames;
+			DebugAssertIndex(charCreationProvinceNames, raceIndex);
+			const std::string &provinceName = charCreationProvinceNames[raceIndex];
+
+			const auto &pluralRaceNames = exeData.races.pluralNames;
+			DebugAssertIndex(pluralRaceNames, raceIndex);
+			const std::string &pluralRaceName = pluralRaceNames[raceIndex];
+
+			const auto &charClassLibrary = game.getCharacterClassLibrary();
+			const int charClassDefID = charCreationState.getClassDefID();
+			const auto &charClassDef = charClassLibrary.getDefinition(charClassDefID);
+
+			// Replace first %s with player class.
+			size_t index = segment.find("%s");
+			segment.replace(index, 2, charClassDef.getName());
+
+			// Replace second %s with player name.
+			index = segment.find("%s");
+			segment.replace(index, 2, charCreationState.getName());
+
+			// Replace third %s with province name.
+			index = segment.find("%s");
+			segment.replace(index, 2, provinceName);
+
+			// Replace fourth %s with plural race name.
+			index = segment.find("%s");
+			segment.replace(index, 2, pluralRaceName);
+
+			// If player is female, replace "his" with "her".
+			if (!charCreationState.isMale())
+			{
+				index = segment.rfind("his");
+				segment.replace(index, 3, "her");
+			}
+
+			return segment;
+		}();
+
+		const int lineSpacing = 1;
+
+		const RichTextString richText(
+			text,
+			FontName::Arena,
+			textColor,
+			TextAlignment::Center,
+			lineSpacing,
+			game.getFontLibrary());
+
+		const int textureHeight = std::max(richText.getDimensions().y, 40);
+		Texture texture = TextureUtils::generate(TextureUtils::PatternType::Parchment,
+			richText.getDimensions().x + 20, textureHeight,
+			game.getTextureManager(), game.getRenderer());
+
+		const Int2 textureCenter(
+			(ArenaRenderUtils::SCREEN_WIDTH / 2) - 1,
+			(ArenaRenderUtils::SCREEN_HEIGHT / 2) - 1);
+
+		return std::make_unique<TextSubPanel>(game, center, richText,
+			toSecondSubPanel, std::move(texture), textureCenter);
+	}();
+
+	game.pushSubPanel(std::move(firstSubPanel));
+}
+
+void CharacterCreationUiController::onChooseRaceProvinceCancelButtonSelected(Game &game)
+{
+	game.popSubPanel();
+
+	// Push the initial text sub-panel.
+	std::unique_ptr<Panel> textSubPanel = ChooseRacePanel::getInitialSubPanel(game);
+	game.pushSubPanel(std::move(textSubPanel));
 }
 
 void CharacterCreationUiController::onBackToRaceSelectionButtonSelected(Game &game)
